feat(rust): extract a cluster repl command for local development
