+++ implementations/rust/ockam/ockam_command/src/cluster/static/repl/long_about.txt
@@ -1,0 +1,1 @@
+This command can be used to access the Repl of an agent started locally.
